所有后端文件以这个为主

+++ backend/src/main/java/com/pbdcompany/controller/ReviewController.java
@@ -37,7 +37,7 @@ public class ReviewController {
     }
     @PostMapping("/{orderId}")
     public ResponseEntity<?> addReview(
-            @PathVariable Long orderId,
+            @PathVariable int orderId,
             @RequestBody ReviewRequest request,
             HttpServletRequest httpServletRequest) {
